Changes; - Sorted out the #includes to now include the .hpp and not use the .cpp file. This is a much better way to do it and stops the issues of only being allowed one include of the .cpp file making it harder to know what includes files are using

+++ Code/GreenOnion/bloom/MurmurHash3.h
@@ -26,7 +26,7 @@ typedef unsigned __int64 uint64_t;
 
 //-----------------------------------------------------------------------------
 
-uint MurmurHash3_x86_32  ( const uint long * key, int len, uint32_t seed);
+uint32_t MurmurHash3_x86_32  ( const uint32_t * key, int len, int seed);
 
 //-----------------------------------------------------------------------------
